Move the toSnake test back to main package per Brian's request
